requirements.txt
pdf2image==1.17.0
langchain~=0.3.4
langchain-openai==0.2.4
langchain-community~=0.3.3
python-dotenv==1.0.1
openai~=1.52.2
matplotlib==3.9.2
pandas~=2.2
tiktoken~=0.8.0
azure-identity~=1.19.0
azure-search-documents==11.5.1
azure-ai-generative==1.0.0b11
azure-cognitiveservices-speech~=1.41.1
azure-ai-documentintelligence==1.0.0b2
scikit-learn==1.5.2
datasets~=3.0.2
opencv-python-headless==4.10.0.84
Markdown==3.7
markdownify==0.13.1
pymupdf==1.24.12
Jinja2==3.1.5
jsonlines==4.0.0
promptflow[azure]==1.16.1
promptflow-tools==1.4.0
azure-ai-contentsafety==1.0.0
einops==0.8.0
onnx>=1.15.0
onnxruntime>=1.18.0
onnxscript>=0.1.0.dev20240126
git+https://github.com/microsoft/Olive
peft
pickleshare==0.7.5
httpx==0.27.0
azure-ai-inference==1.0.0b6
trl
datasets
pyOpenSSL
tf-keras
bitsandbytes>=0.39.0
accelerate>=0.20.0
huggingface_hub
sentencepiece
requests
deepspeed
unsloth
transformers~=4.48.2